switch to template install

+++ src/build.py
@@ -6,7 +6,11 @@ from . import file_util
 from .build_args import BuildArgs
 from .processing.token import Token
 from .processing.packages import Packages
+from .processing.req_packages import ReqPackages
 from .processing.update import Update
+from .processing.json_config import JsonConfig
+from .processing.default_resource import DefaultResource
+from .install.pre_install_pure import PreInstallPure
 
 
 class Build:
@@ -14,7 +18,7 @@ class Build:
 
     def __init__(self, args: BuildArgs) -> None:
         self._config = Config()
-        self._build_path = self._config.root_path / self._config.build_dir_name
+        self._build_path = self._config.build_path
         self._args = args
         self._src_path = self._config.root_path / args.oxt_src
         if not self._src_path.exists():
@@ -24,15 +28,23 @@ class Build:
 
     def build(self) -> None:
         """Builds the project."""
+        # sourcery skip: extract-method
         if self._args.clean:
             self.clean()
         # self._ensure_build()
         self._copy_src_dest()
+        self._rename_lo_pip()
         if self._args.process_tokens:
             self._process_tokens()
 
+        self._process_config()
+        self._copy_py_req_packages()
+        self._copy_py_req_files()
+        self._clear_req_cache()
+        self._zip_req_python_path()
+
         if self._args.process_py_packages:
-            pythonpath = self._build_path / "pythonpath"
+            pythonpath = self._build_path / self._config.py_pkg_dir
             if pythonpath.exists():
                 shutil.rmtree(pythonpath)
 
@@ -40,6 +52,11 @@ class Build:
             self._copy_py_files()
             self._clear_cache()
             self._zip_python_path()
+
+        if self._args.pre_install_pure_packages:
+            self._pre_install_pure_packages()
+
+        self._ensure_default_resource()
 
         if self._args.make_dist:
             self._zip_build()
@@ -55,10 +72,24 @@ class Build:
         if self._build_path.exists():
             shutil.rmtree(self._build_path)
 
+    def _rename_lo_pip(self) -> None:
+        """Renames the lo_pip folder."""
+        token = Token()
+        src_dir = self._build_path / "___lo_pip___"
+        if not src_dir.exists():
+            raise FileNotFoundError(f"lo_pip folder '{src_dir}' not found")
+        dest_dir = self._build_path / token.get_token_value("lo_pip")
+        os.rename(src_dir, dest_dir)
+
     def _ensure_build(self) -> None:
         """Ensures the build directory exists."""
         if not self._build_path.exists():
             self._build_path.mkdir(parents=True, exist_ok=True)
+
+    def _ensure_default_resource(self) -> None:
+        """Ensures the default resource file exists."""
+        default_resource = DefaultResource()
+        default_resource.ensure_default()
 
     def _copy_src_dest(self) -> None:
         """Copies the source files to the build directory."""
@@ -71,25 +102,70 @@ class Build:
             text = file_util.read_file(file)
             text = self.process_tokens(text)
             file_util.write_string_to_file(file, text)
+        # process py_runner.py
+        file = self._build_path / "py_runner.py"
+        if file.exists():
+            py_file = str(file)
+            text = file_util.read_file(py_file)
+            text = self.process_tokens(text)
+            file_util.write_string_to_file(py_file, text)
+
+    def _process_config(self) -> None:
+        token = Token()
+        config_file = self._build_path / token.get_token_value("lo_pip") / "config.json"
+        json_config = JsonConfig()
+        json_config.update_json_config(config_file)
 
     def _copy_py_packages(self) -> None:
         """Copies the python packages to the build directory."""
         packages = Packages()
-        packages.copy_packages(self._build_path / "pythonpath")
+        packages.copy_packages(self._build_path / self._config.py_pkg_dir)
 
     def _copy_py_files(self) -> None:
         """Copies the python files to the build directory."""
         packages = Packages()
-        packages.copy_files(self._build_path / "pythonpath")
+        packages.copy_files(self._build_path / self._config.py_pkg_dir)
+
+    def _copy_py_req_packages(self) -> None:
+        """Copies the python packages to the build directory."""
+        packages = ReqPackages()
+        packages.copy_packages(self._build_path / f"req_{self._config.py_pkg_dir}")
+
+    def _copy_py_req_files(self) -> None:
+        """Copies the python files to the build directory."""
+        packages = ReqPackages()
+        packages.copy_files(self._build_path / f"req_{self._config.py_pkg_dir}")
 
     def _clear_cache(self) -> None:
         """Cleans the cache."""
         packages = Packages()
-        packages.clear_cache(self._build_path / "pythonpath")
+        if not packages.has_modules():
+            return
+        packages.clear_cache(self._build_path / self._config.py_pkg_dir)
+
+    def _clear_req_cache(self) -> None:
+        """Cleans the cache."""
+        packages = ReqPackages()
+        packages.clear_cache(self._build_path / f"req_{self._config.py_pkg_dir}")
 
     def _zip_python_path(self) -> None:
         """Zips the python path."""
-        pth = self._build_path / "pythonpath"
+        pth = self._build_path / self._config.py_pkg_dir
+        if not pth.exists():
+            return
+        file_util.zip_folder(folder=pth)
+        shutil.rmtree(pth)
+
+    def _pre_install_pure_packages(self) -> None:
+        """Installs the pure python packages."""
+        pre_install = PreInstallPure()
+        pre_install.install()
+
+    def _zip_req_python_path(self) -> None:
+        """Zips the required packages path."""
+        pth = self._build_path / f"req_{self._config.py_pkg_dir}"
+        if not pth.exists():
+            return
         file_util.zip_folder(folder=pth)
         shutil.rmtree(pth)
